Media bug fixes
