+++ Slide for Reddit/MainViewController.swift
@@ -620,7 +620,6 @@ class MainViewController: ColorMuxPagingViewController, UIPageViewControllerData
             }
         }
         
-        makeMenuNav()
         doButtons()
         var wasntHidden = false
         if !menuNav!.view.isHidden {
@@ -940,19 +939,6 @@ class MainViewController: ColorMuxPagingViewController, UIPageViewControllerData
         UIApplication.shared.statusBarView?.backgroundColor = .clear
         
         menuNav?.view.isHidden = true
-        
-        if let session = (UIApplication.shared.delegate as? AppDelegate)?.session {
-            if AccountController.isLoggedIn && AccountController.isGold {
-                do {
-                    try session.setVisited(names: History.currentSeen) { (result) in
-                        
-                    }
-                } catch {
-                    
-                }
-            }
-            History.currentSeen.removeAll()
-        }
     }
 
     func showSortMenu(_ sender: UIButton?) {
