+++ Slide for Reddit/ModalMediaViewController.swift
@@ -7,24 +7,12 @@
 //
 
 import Anchorage
-import SubtleVolume
 import Then
 import UIKit
 
 class ModalMediaViewController: UIViewController {
 
 //    var loadedURL: URL?
-
-    let volume = SubtleVolume(style: SubtleVolumeStyle.roundedLine)
-    let volumeHeight: CGFloat = 3
-    
-    var safeAreaInsets: UIEdgeInsets {
-        if #available(iOS 11.0, tvOS 11.0, *) {
-            return view.safeAreaInsets
-        } else {
-            return UIEdgeInsets.zero
-        }
-    }
 
     var embeddedVC: EmbeddableMediaViewController!
     var fullscreen = false
@@ -244,24 +232,6 @@ class ModalMediaViewController: UIViewController {
         } else {
             shouldLoad = true
         }
-        
-        volume.barTintColor = .white
-        volume.barBackgroundColor = UIColor.white.withAlphaComponent(0.3)
-        volume.animation = .slideDown
-        view.addSubview(volume)
-        
-        NotificationCenter.default.addObserver(volume, selector: #selector(SubtleVolume.resume), name: NSNotification.Name.UIApplicationDidBecomeActive, object: nil)
-    }
-
-    override func viewDidLayoutSubviews() {
-        layoutVolume()
-    }
-    
-    func layoutVolume() {
-        let volumeYPadding: CGFloat = 10
-        let volumeXPadding = UIScreen.main.bounds.width * 0.4 / 2
-        volume.superview?.bringSubview(toFront: volume)
-        volume.frame = CGRect(x: safeAreaInsets.left + volumeXPadding, y: safeAreaInsets.top + volumeYPadding, width: UIScreen.main.bounds.width - (volumeXPadding * 2) - safeAreaInsets.left - safeAreaInsets.right, height: volumeHeight)
     }
 
     override func viewWillAppear(_ animated: Bool) {
